+++ test/explicit/documentation/output/swagger_test.rb
@@ -24,34 +24,23 @@ class Explicit::Documentation::Output::SwaggerTest < ActiveSupport::TestCase
 
   test "POST /registrations" do
     swagger.dig(:paths, "/registrations", "post").tap do |req|
-      assert_equal ["Auth"], req.dig(:tags)
+      assert_equal [ "Auth" ], req.dig(:tags)
       assert_equal "Registration", req.dig(:summary)
       assert_match "Attempts to register a new user", req.dig(:description)
       assert_nil req.dig(:parameters)
 
       body_schema = req.dig(:requestBody, :content, "application/json", :schema)
       assert_equal "object", body_schema[:type]
-      assert_equal ["name", "email_address", "password", "terms_of_use"], body_schema[:required]
+      assert_equal [ "name", "email_address", "password", "terms_of_use" ], body_schema[:required]
     end
   end
 
   test "GET /articles/:article_id" do
     swagger.dig(:paths, "/articles/{article_id}", "get").tap do |req|
-      assert_equal ["Articles"], req.dig(:tags)
+      assert_equal [ "Articles" ], req.dig(:tags)
       assert_equal "Get article", req.dig(:summary)
-      
-      assert_equal req.dig(:parameters, 0), {
-        name: "Authorization",
-        in: "header",
-        required: true,
-        schema: {
-          type: "string",
-          pattern: "Bearer [a-zA-Z0-9]{20}"
-        },
-        style: "simple"
-      }
 
-      assert_equal req.dig(:parameters, 1), {
+      assert_equal req.dig(:parameters, 0), {
         name: "article_id",
         in: "path",
         required: true,
@@ -62,15 +51,17 @@ class Explicit::Documentation::Output::SwaggerTest < ActiveSupport::TestCase
         style: "simple"
       }
 
+      assert_equal req.dig(:security), { bearerAuth: [] }
+
       assert_nil req.dig(:requestBody)
     end
   end
 
   test "default value serialization" do
-    type = type([:default, 10, :integer])
+    type = type([ :default, 10, :integer ])
     assert_equal 10, type.swagger_schema[:default]
 
-    type = type([:default, -> { 10 }, :integer])
+    type = type([ :default, -> { 10 }, :integer ])
     assert_nil type.swagger_schema[:default]
   end